bevy: basic movement animation
So far, pieces moved instantly to their new position. We use the fact
that the position as considered by the chess logic is stored in the
`PieceMarker` component. We now do not update the rendering position
when the logical position changes. Instead we introduce a system that
gradually moves pieces from their rendering to their logical position.

The change does not affect the logic and input, the opponent can make a
move while a piece is still in motion.

The animation looks great for sliding movements, but the knight can clip
through other pieces and also with diagonal movements collisions with
nearby pieces are possible.