bevy: rework input
Currently, the design is focussed around mouse input. A system watches
the mouse movements and already converts to board coordinates. On
left-mouse-button click, we check where the mouse is currently on the
board and perform the logic.

The rework simplifies the architecture and always for simple integration
of touch inputs. One system watches the mouse buttons and fires a
RawClickEvent when clicked. An observer then performs the board ray cast
and fires a BoardClickEvent. The selection/move logic can now be an
observer of the BoardClickEvent.